schemacrawler_pytest_poc/requirements.txt
pyyaml
pytest>=8.0
